api/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.10.4
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
redis==4.5.5
prometheus-client==0.16.0
python-multipart==0.0.6
httpx==0.24.0
pydantic-settings==2.0.2
tenacity==8.2.2
